+++ env_canada/__init__.py
@@ -4,6 +4,7 @@ __all__ = [
     "ECHistoricalRange",
     "ECHydro",
     "ECRadar",
+    "ECMap",
     "ECWeather",
     "ECWeatherUpdateFailed",
 ]
@@ -12,4 +13,5 @@ from .ec_aqhi import ECAirQuality
 from .ec_historical import ECHistorical, ECHistoricalRange
 from .ec_hydro import ECHydro
 from .ec_radar import ECRadar
+from .ec_map import ECMap
 from .ec_weather import ECWeather, ECWeatherUpdateFailed